[v2] Get rid of deprecated types
