+++ src/schema/v2/schema.ts
@@ -44,7 +44,6 @@ import Partner from "./partner"
 import Partners from "./partners"
 import FilterPartners from "./filter_partners"
 import filterArtworks from "./filter_artworks"
-import FilterSaleArtworks from "./filter_sale_artworks"
 import FollowArtist from "./me/follow_artist"
 import FollowProfile from "./me/follow_profile"
 import FollowGene from "./me/follow_gene"
@@ -129,7 +128,6 @@ const rootFields: GraphQLFieldConfigMap<any, ResolverContext> = {
   // FIXME: Expected 1 arguments, but got 0
   // @ts-ignore
   filter_artworks: filterArtworks(),
-  filter_sale_artworks: FilterSaleArtworks,
   gene: Gene,
   genes: Genes,
   suggested_genes: SuggestedGenes,
